+++ Assets/CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [0.4.4] - 2023-06-23
+### Added
+- Added window ClearUpLocalization for Scenes and Prefabs
+- Added tips for search text in scenes or prefabs
+### Changed
+- Renamed window Auto Translate for Unity Localization
+### Fixed
+- Fixed swap word into key when translate smart word
+- Fixed when code key to replace string key in smartword
+
 ## [0.4.3] - 2023-06-22
 ### Added
 - Added mapping dialects of languages for google free api